R3c: the three-arm W4 verify gate, docs, and the default flip
The gate the flipped default never had on main (tag pre-port-20260829, doc
56 H2b): verify_adapter build_hf_model (dequant-patched oracle) + the
LLMS_VERIFY_{PREFILL,DECODE}_CACHE / _PREFILL_M load-don't-compile
contract; run_npu2_verify.lit three arms (bf16 A/B pinned =0, w4 vs the
patched oracle pinned =1 loading arm 1's prefill set, the quantization bar
vs the UNPATCHED checkpoint loading both) sharing ONE artifact set;
Makefile verify-quant-bar + clean-build (clean deletes shared
verify/reports on purpose-noted grounds); README/ARCHITECTURE structural
lines (perf attributed to the study record or dropped; llms README = one
tree line, operator rule Q16). verify_runner re-gains LLMS_VERIFY_MAX_SEQ
(dropped from #33 by its review) and the report now RECORDS max_seq --
declared in the PR body. W4_DEFAULT flips True in the same PR as its gate
(#32/#33 adjudications held it OFF until this landed); the pinned default
asserts in both test files flip with it.

Host gates: test_w4_decode_pack 4/4, test_w4_decode_driver 4/4 at this
head (agents/.state/r3c/evidence/{pack,driver}-tests-flip.log), adapter
imports, black clean. Device: arms 1/2 green pre-run at the #33 head +
this adapter (devq 885/886: fresh compile, LOADED set, patched-oracle
banner, topk 2/2 PASS each, Turbo); arm 3 quant-bar, the qwen3_1_7b
regression, and the mis-packed red-before are devq 887-889 (results on
the PR before open).

+++ programming_examples/llms/qwen3_0_6b/test_w4_decode_driver.py
@@ -41,7 +41,7 @@ def _decode_at(flag):
 def test_selected_artifact_set_follows_the_flag():
     bf16 = ("rms_qkv_qknorm_rope_gemv2", "o_gemv_ffn", "lm_head_gemv")
     int4 = ("rms_qkv_qknorm_rope_gemv2", "o_gemv_ffn_int4", "lm_head_gemv")
-    assert _decode_at(None).required_decode_artifacts() == bf16  # default OFF
+    assert _decode_at(None).required_decode_artifacts() == int4  # default ON
     assert _decode_at("0").required_decode_artifacts() == bf16
     assert _decode_at("1").required_decode_artifacts() == int4
